rename to GoogleCloudTraceContext
