--- propagator/google_cloud_platform/CHANGELOG.md
@@ -1,1 +1,0 @@
-# Release History: opentelemetry-propagator-google_cloud_platform
