+++ propagator/google_cloud_trace_context/CHANGELOG.md
@@ -1,0 +1,1 @@
+# Release History: opentelemetry-propagator-google_cloud_trace_context
