+++ propagator/google_cloud_trace_context/lib/opentelemetry-propagator-google_cloud_trace_context.rb
@@ -5,4 +5,4 @@
 # SPDX-License-Identifier: Apache-2.0
 
 require 'opentelemetry-api'
-require_relative 'opentelemetry/propagator/google_cloud_platform'
+require_relative 'opentelemetry/propagator/google_cloud_trace_context'
