+++ propagator/google_cloud_trace_context/lib/opentelemetry/propagator/google_cloud_trace_context.rb
@@ -4,6 +4,9 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+require 'opentelemetry/propagator/google_cloud_trace_context/version'
+require 'opentelemetry/propagator/google_cloud_trace_context/text_map_propagator'
+
 # OpenTelemetry is an open source observability framework, providing a
 # general-purpose API, SDK, and related tools required for the instrumentation
 # of cloud-native software, frameworks, and libraries.
@@ -13,9 +16,19 @@
 module OpenTelemetry
   # Namespace for OpenTelemetry propagator extension libraries
   module Propagator
-    # Namespace for OpenTelemetry GoogleCloudPlatform propagation
-    module GoogleCloudPlatform
-      VERSION = '0.0.0'
+    # Namespace for OpenTelemetry GoogleCloudTraceContext propagation
+    module GoogleCloudTraceContext
+      extend self
+
+      TEXT_MAP_PROPAGATOR = TextMapPropagator.new
+
+      private_constant :TEXT_MAP_PROPAGATOR
+
+      # Returns a text map propagator that propagates context using the
+      # GoogleCloudTraceContext format.
+      def text_map_propagator
+        TEXT_MAP_PROPAGATOR
+      end
     end
   end
 end
