+++ propagator/google_cloud_trace_context/test/test_helper.rb
@@ -8,7 +8,7 @@ require 'simplecov'
 require 'bundler/setup'
 Bundler.require(:default, :development, :test)
 
-require 'opentelemetry-propagator-google_cloud_platform'
+require 'opentelemetry-propagator-google_cloud_trace_context'
 require 'minitest/autorun'
 
 OpenTelemetry.logger = Logger.new($stderr, level: ENV.fetch('OTEL_LOG_LEVEL', 'fatal').to_sym)
