+++ release-please-config.json
@@ -210,9 +210,9 @@
       "package-name": "opentelemetry-propagator-xray",
       "version-file": "lib/opentelemetry/propagator/xray/version.rb"
     },
-    "propagator/google_cloud_platform": {
-      "package-name": "opentelemetry-propagator-google_cloud_platform",
-      "version-file": "lib/opentelemetry/propagator/google_cloud_platform/version.rb"
+    "propagator/google_cloud_trace_context": {
+      "package-name": "opentelemetry-propagator-google_cloud_trace_context",
+      "version-file": "lib/opentelemetry/propagator/google_cloud_trace_context/version.rb"
     },
     "resources/azure": {
       "package-name": "opentelemetry-resource-detector-azure",
@@ -222,9 +222,9 @@
       "package-name": "opentelemetry-resource-detector-container",
       "version-file": "lib/opentelemetry/resource/detector/container/version.rb"
     },
-    "resources/google_cloud_platform": {
-      "package-name": "opentelemetry-resource-detector-google_cloud_platform",
-      "version-file": "lib/opentelemetry/resource/detector/google_cloud_platform/version.rb"
+    "resources/google_cloud_trace_context": {
+      "package-name": "opentelemetry-resource-detector-google_cloud_trace_context",
+      "version-file": "lib/opentelemetry/resource/detector/google_cloud_trace_context/version.rb"
     }
   }
 }
